project_19_sql_querying_agent/requirements.txt
langchain
langchain-google-genai
python-dotenv
langchain-community
sqlalchemy
uv